# bruno/cms-api-openapi.yaml
openapi: 3.0.3
info:
  title: Complaint Management System API
  version: "1.0.0"
  description: Bruno-importable full API collection for complaint-management-api.
servers:
  - url: https://localhost:7000
    description: Local HTTPS
  - url: http://localhost:5000
    description: Local HTTP
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: Search users
      operationId: searchUsers
      parameters:
        - $ref: '#/components/parameters/Keyword'
        - $ref: '#/components/parameters/Role'
        - $ref: '#/components/parameters/IsActive'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { description: OK }
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201': { description: Created }
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': { description: OK }
        '404': { description: Not found }
    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200': { description: Updated }
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': { description: Deleted }
...
  /api/v1/users/me/change-password:
    put:
      tags: [Users]
      summary: Change current user password
      operationId: changeMyPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200': { description: Password changed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    Role:
      name: role
      in: query
      schema:
        type: string
    IsActive:
      name: isActive
      in: query
      schema:
        type: boolean
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        default: 20
  schemas:
...
    CreateUserRequest:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        username: { type: string }
        phoneNumber: { type: string }
        password: { type: string }
        role: { type: string, example: Agent }
      required: [name, email, username, phoneNumber, password, role]
    UpdateUserRequest:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        username: { type: string }
        phoneNumber: { type: string }
        role: { type: string }
      required: [name, email, username, phoneNumber, role]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
        confirmPassword: { type: string }
      required: [currentPassword, newPassword, confirmPassword]
...